feat:filter by individual fields
