+++ src/api/domain/farmer/controller.py
@@ -1,8 +1,6 @@
 import api.domain.farmer.repository as Repository
 from flask import jsonify,request
-from api.models.index import Farmer
-
-
+from api.models.index import Farmer, Technician
 
 ## Controller get farmer by user_id
 def get_farmer_by_user_owner(id):    
@@ -11,4 +9,15 @@ def get_farmer_by_user_owner(id):
         return jsonify('No existe el granjero')   
    return farmer.serialize()
    
-    
+## FILTER TECH
+def filter_tech(body):
+    if body["name"]:
+        print(body["name"])  
+        tech = Repository.filter_tech_by_name(body["name"])
+        return tech
+    if body["ccaa"]:
+       tech = Repository.filter_tech_by_ccaa(body["ccaa"])
+       return tech
+    if body["speciality"]:
+        tech = Repository.filter_tech_by_speciality(body["speciality"])
+        return tech
