+++ src/api/domain/farmer/router.py
@@ -12,4 +12,10 @@ api = Blueprint("api/farmer", __name__)
 def get_farmer(id):
     farmer = Controller.get_farmer_by_user_owner(id)       
     return jsonify(farmer), 200
-
+    
+# FILTER TECH
+@api.route('/', methods=['POST'])
+def filter_tech():
+    body = request.get_json()
+    filtered_tech = Controller.filter_tech(body)
+    return jsonify(filtered_tech), 200
