+++ src/api/domain/technician/controller.py
@@ -22,14 +22,3 @@ def get_tech_by_user_owner(user_id):
         return jsonify('No existe ese técnico')
     return tech.serialize()
 
-## FILTER TECH
-def filter_tech(body):
-    if body["name"]:
-        tech = Repository.filter_tech_by_name(body["name"])
-        return tech
-    if body["ccaa"]:
-        tech = Repository.filter_tech_by_ccaa(body["name"])
-        return tech
-    if body["speciality"]:
-        tech = Repository.filter_tech_by_speciality(body["speciality"])
-        return tech
