+++ src/api/domain/technician/repository.py
@@ -19,22 +19,6 @@ def get_all_tech():
 ### GET TECH BY USER_OWNER
 def get_tech_by_user_owner(user_id):
     return Technician.query.filter_by(user_owner=user_id).first()
-#########################################################################
-## FILTER TECH BY NAME
- def filter_tech_by_name(name):
-    tech = Technician.query.filter_by(name=name).all()
-    tech_serialized = list(map(lambda x : x.serialize()), tech)
-    return tech_serialized
 
-## FILTER TECH BY CCAA
-def filter_tech_by_ccaa(ccaa):
-    tech = Technician.query.filter_by(ccaa=ccaa).all()
-    tech_serialized = list(map(lambda x : x.serialize()), tech)
-    return tech_serialized
 
-## FILTER TECH BY SPECIALITY
-def filter_tech_by_speciality(speciality):
-    tech = Technician.query.filter_by(speciality=speciality).all()
-    tech_serialized = list(map(lambda x : x.serialize()), tech)
-    return tech_serialized
