+++ src/api/domain/technician/router.py
@@ -16,5 +16,3 @@ def get_all_tech():
 def get_one_tech(id):
     one_tech = Controller.get_tech_by_user_owner(id)
     return jsonify(one_tech), 200
-    
-
